test: replace mocha with hydro
