--- test/support/bootstrap.js
@@ -1,18 +1,0 @@
-
-/*!
- * External dependencies.
- */
-
-var chai = require('chai');
-
-/*!
- * Register `should`.
- */
-
-global.should = chai.should();
-
-/*!
- * Include stack traces.
- */
-
-chai.Assertion.includeStack = true;
